# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quire Server API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "description": "Private reading-data sync and EPUB storage. Account and watched-folder administration are owner CLI operations."
  },
  "paths": {
...
    "/v1/admin/libraries/{library}/books": {
      "get": {
        "summary": "Get admin libraries {library} books",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/admin/libraries/{library}/books/{id}/metadata": {
      "put": {
        "summary": "Put admin libraries {library} books {id} metadata",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/libraries": {
      "get": {
        "summary": "Get libraries",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        }
      }
    },
    "/v1/admin/libraries": {
      "get": {
        "summary": "Get admin libraries",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      },
      "post": {
        "summary": "Post admin libraries",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          },
          "201": {
            "description": "Resource created."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      }
    },
    "/v1/admin/libraries/{id}/members/{user}": {
      "put": {
        "summary": "Put admin libraries {id} members {user}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "summary": "Delete admin libraries {id} members {user}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/admin/invites/{id}/libraries/{library}": {
      "put": {
        "summary": "Put admin invites {id} libraries {library}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/admin/watches": {
      "get": {
        "summary": "Get admin watches",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      },
      "post": {
        "summary": "Post admin watches",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          },
          "201": {
            "description": "Resource created."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      }
    },
    "/v1/admin/watches/{id}": {
      "delete": {
        "summary": "Delete admin watches {id}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/admin/watches/{id}/scan": {
      "post": {
        "summary": "Post admin watches {id} scan",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          },
          "201": {
            "description": "Resource created."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/admin/overview": {
      "get": {
        "summary": "Get admin overview",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions."
      }
    },
    "/v1/admin/libraries/{library}/books/{id}": {
      "put": {
        "summary": "Put admin libraries {library} books {id}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "summary": "Delete admin libraries {library} books {id}",
        "tags": [
          "Administration"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed; GET returns the requested representation."
          },
          "204": {
            "description": "Mutation completed without a response body."
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Administrator permission required."
          },
          "409": {
            "description": "Conflicting state."
          }
        },
        "description": "Requires an active administrator account. Does not grant access to another user's annotations or reading positions.",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}